requirements.txt

# Core dependencies
tkinter
pillow>=9.0.0
pystray>=0.19.0

# For global hotkeys
pynput>=1.7.6

# For system theme detection (optional)
darkdetect

# For development and testing
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
isort>=5.12.0
mypy>=1.0.0
flake8>=6.0.0
